Extracting deposit overview utilities
Now moved to a separate `client/deposits/utils/index.js` file. it's not top-level placement, but the path makes a lot of sense to me.
